% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_completo}
\alias{dataset_completo}
\title{Dataset Completo de Condiciones Meteorologicas Históricas}
\format{
A data frame with X rows and 24 variables:
\describe{
\item{id}{Character. Identificador unico de la estación meteorologica.}
\item{fecha}{Date. La fecha de la medicion en formato \code{AAAA-MM-DD}.}
\item{temperatura_abrigo_150cm}{Numeric. Temperatura medida a 150 cm sobre el suelo.}
\item{temperatura_abrigo_150cm_maxima}{Numeric. Temperatura maxima registrada a 150 cm sobre el suelo.}
\item{temperatura_abrigo_150cm_minima}{Numeric. Temperatura minima registrada a 150 cm sobre el suelo.}
\item{temperatura_intemperie_5cm_minima}{Numeric. Temperatura minima registrada a 5 cm sobre el suelo.}
\item{temperatura_intemperie_50cm_minima}{Numeric. Temperatura mínima registrada a 50 cm sobre el suelo.}
\item{temperatura_suelo_5cm_media}{Numeric. Temperatura media del suelo a 5 cm de profundidad.}
\item{temperatura_suelo_10cm_media}{Numeric. Temperatura media del suelo a 10 cm de profundidad.}
\item{temperatura_inte_5cm}{Numeric. Temperatura a 5 cm sobre el nivel del suelo.}
\item{temperatura_intemperie_150cm_minima}{Numeric. Temperatura mínima a 150 cm sobre el nivel del suelo.}
\item{humedad_suelo}{Numeric. Humedad del suelo.}
\item{precipitacion_pluviometrica}{Numeric. Precipitación en milímetros registrada durante el día.}
\item{granizo}{Numeric. Indicador binario (0 o 1) que indica si hubo granizo.}
\item{nieve}{Numeric. Indicador binario (0 o 1) que indica si hubo nieve.}
\item{heliofania_efectiva}{Numeric. Horas de radiación solar efectiva recibida durante el día.}
\item{heliofania_relativa}{Numeric. Proporción de radiación solar relativa recibida.}
\item{tesion_vapor_media}{Numeric. Tensión de vapor media en el aire durante el día.}
\item{humedad_media}{Numeric. Humedad media del aire durante el dia.}
\item{rocio_medio}{Numeric. Temperatura media del punto de rocio.}
\item{duracion_follaje_mojado}{Numeric. Duración del follaje mojado.}
\item{velocidad_viento_200cm_media}{Numeric. Velocidad media del viento registrada a 200 cm sobre el suelo.}
\item{direccion_viento_200cm}{Character. Dirección del viento registrada a 200 cm sobre el suelo.}
\item{velocidad_viento_1000cm_media}{Numeric. Velocidad media del viento registrada a 1000 cm sobre el suelo.}
\item{direccion_viento_1000cm}{Character. Dirección del viento registrada a 1000 cm sobre el suelo.}
\item{presion_media}{Numeric. La presión atmosférica media registrada durante el día.}
\item{radiacion_global}{Numeric. Radiación solar global recibida durante el día, en unidades de W/m².}
\item{radiacion_neta}{Numeric. Radiación neta recibida durante el día.}
\item{evaporacion_tanque}{Numeric. Medición de la evaporación registrada en un tanque durante el día.}
\item{evapotranspiracion_potencial}{Numeric. Medición de la evapotranspiración potencial.}
\item{profundidad_napa}{Numeric. Profundidad de la napa freática.}
\item{horas_frio}{Numeric. Número total de horas de frío registrado durante el día.}
\item{unidad_frio}{Numeric. Unidad de medición para las horas de frío.}
}
}
\source{
datos meteorologicos obtenidos por mis profesoras :D
}
\usage{
dataset_completo
}
\description{
Este dataset contiene mediciones diarias de varias variables meteorologicas de diversas estaciones.
Las columnas incluyen datos sobre temperaturas, precipitaciones, humedad, viento, radiacion solar y otros factores climaticos.
El dataset es útil para análisis climáticos históricos y predicciones basadas en condiciones pasadas.
}
\keyword{datasets}
