% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_precipitacion.R
\name{grafico_precipitacion}
\alias{grafico_precipitacion}
\title{Crear grafico de precipitacion pluviometrica}
\usage{
grafico_precipitacion(dataset, anio)
}
\arguments{
\item{dataset}{Un dataframe con los datos meteorologicos historicos}

\item{anio}{Un valor numerico o caracter correspondiente al año para el cual se desea obtener el grafico}
}
\value{
esta funcion devuelve un grafico con la precipitacion pluviometrica anual
}
\description{
Crear grafico de precipitacion pluviometrica
}
\examples{
grafico_precipitacion(dataset_completo, 2010)

}
