% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guia_de_supervivencia.R
\name{guia_de_supervivencia}
\alias{guia_de_supervivencia}
\title{Guia de supervivencia para viajeros en el tiempo}
\usage{
guia_de_supervivencia(fecha)
}
\arguments{
\item{fecha}{el usuario debe ingresar la fecha a la cual quiere viajar}
}
\value{
te devuelve un outfit en base a la temperatura del dia
}
\description{
Guia de supervivencia para viajeros en el tiempo
}
\examples{
guia_de_supervivencia("2011-04-25")

}
