% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectura_datos.R
\name{lectura_datos}
\alias{lectura_datos}
\title{Lectura Datos}
\usage{
lectura_datos(id_estacion)
}
\arguments{
\item{id_estacion}{este es el id de la estacion}
}
\value{
esto devuelve los datasets por estacion
}
\description{
Lectura Datos
}
\examples{
lectura_datos ("NH0046")
lectura_datos ("NH0098")
lectura_datos ("NH0437")
lectura_datos ("NH0472")
lectura_datos ("NH0910")

}
