% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumen_meteorologico.R
\name{resumen_meteorologico}
\alias{resumen_meteorologico}
\title{Resumen meteorologico anual}
\usage{
resumen_meteorologico(dataset, anio_usuario)
}
\arguments{
\item{dataset}{Un dataframe con los datos meteorologicos historicos}

\item{anio_usuario}{Un valor numerico o caracter correspondiente al anio para el cual se desea obtener el resumen}
}
\value{
esta funcion devuelve diferentes valores con diferentes fenomenos meteorologicos anuales
}
\description{
Resumen meteorologico anual
}
\examples{
resumen_meteorologico(dataset_completo, anio_usuario = 1973)

}
