% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unir_datasets.R
\name{unir_datasets}
\alias{unir_datasets}
\title{Unir datasets}
\usage{
unir_datasets(ids_estaciones)
}
\arguments{
\item{ids_estaciones}{Un vector de caracteres con los IDs de las estaciones meteorologicas
que se desean combinar. Los IDs se utilizan para acceder a los archivos correspondientes
a cada estacion.}
}
\value{
devuelve los datasets unidos
}
\description{
Esta funcion toma una lista de IDs de estaciones meteorologicas y los usa para leer
y combinar los datos historicos correspondientes a cada estacion en un solo dataframe.
El resultado es un dataframe que contiene los datos combinados de todas las estaciones.
}
\examples{
#' Definir los IDs de las estaciones
ids_estaciones <- c("NH0046", "NH0098", "NH0437", "NH0472", "NH0910")
#Llamar a la funcion para unir los datasets
dataset_completo <- unir_datasets(ids_estaciones)

}
