% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viajando_en_el_tiempo.R
\name{viajando_en_el_tiempo}
\alias{viajando_en_el_tiempo}
\title{Viajando en el tiempo}
\usage{
viajando_en_el_tiempo(fecha_nacimiento, fecha_viaje)
}
\arguments{
\item{fecha_nacimiento}{ingresar fecha de nacimiento en formato aaaa/mm/dd}

\item{fecha_viaje}{ingresar fecha de viaje en el tiempo en formato aaaa/mm/dd}
}
\value{
esta funcion te devuelve un mensaje para ti viaje en base a tu signo
}
\description{
Viajando en el tiempo
}
\examples{
viajando_en_el_tiempo("2004-02-09", "1974-12-30")

}
