% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiaco.R
\name{zodiaco}
\alias{zodiaco}
\title{Zodiaco}
\usage{
zodiaco(fecha_nacimiento)
}
\arguments{
\item{fecha_nacimiento}{tu fecha de nacimiento}
}
\value{
te devuelve un outfit segun tu zodiaco
}
\description{
Zodiaco
}
\examples{
ejemplo de uso zodiaco("1990-08-15")


}
