% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conversion}
\alias{conversion}
\title{Conversion de grados}
\usage{
conversion(fahrenheit)
}
\arguments{
\item{fahrenheit}{temperatura en fahrenheit}
}
\value{
devuelve la temperatura en centigrados
}
\description{
Conversion de grados
}
\examples{
conversion(98)

}
