% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_precipitacion.R
\name{grafico_precipitacion}
\alias{grafico_precipitacion}
\title{Crear grafico de precipitacion pluviometrica}
\usage{
grafico_precipitacion(dataset, anio)
}
\arguments{
\item{dataset}{Un dataframe con los datos meteorologicos historicos}

\item{anio}{Un valor numerico o caracter correspondiente al año para el cual se desea obtener el grafico}
}
\value{
Esta funcion devuelve un grafico de barras donde el eje x muestra las fechas de los dias con precipitacion,
y el eje y muestra el numero de dias con precipitacion.
}
\description{
Esta funcion genera un grafico de barras que muestra la cantidad de dias con precipitacion pluviometrica en un ano especifico.
El grafico se basa en los datos de precipitacion almacenados en el dataframe proporcionado, filtrando los datos para el
ano solicitado y solo mostrando los dias con precipitacion mayor a 0.
}
\examples{
grafico_precipitacion(dataset_completo, 2010)

}
