% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectura_datos.R
\name{lectura_datos}
\alias{lectura_datos}
\title{Lectura Datos}
\usage{
lectura_datos(id_estacion)
}
\arguments{
\item{id_estacion}{Un valor de tipo caracter que corresponde al ID de la estacion meteorologica
de la cual se desea obtener los datos. Este ID se utiliza para formar la URL desde donde se
descargan los datos.}
}
\value{
Devuelve un dataframe con los datos meteorologicos de la estacion correspondiente al ID ingresado.
Si los datos no estan disponibles localmente, la funcion los descarga desde una URL predefinida.
}
\description{
Esta funcion permite leer los datos meteorologicos historicos de una estacion especifica
a partir de su ID. Si los datos ya se han descargado previamente, la funcion los lee desde
el directorio temporal; si no, los descarga desde la URL correspondiente y los lee.
}
\examples{
lectura_datos ("NH0046")
lectura_datos ("NH0098")
lectura_datos ("NH0437")
lectura_datos ("NH0472")
lectura_datos ("NH0910")

}
