% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperaturas.R
\name{temperaturas}
\alias{temperaturas}
\title{Temperaturas anuales}
\usage{
temperaturas(dataset, anio_usuario)
}
\arguments{
\item{dataset}{Un dataframe con los datos meteorologicos historicos}

\item{anio_usuario}{Un valor numerico o caracter correspondiente al anio para el cual se desea obtener
las diferentes temperaturas anuales}
}
\value{
esta funcion devuelve el dia mas calido del anio con su temperatura maxima
junto al dia mas frio del anio con su temperatura minima
}
\description{
Temperaturas anuales
}
\examples{
temperaturas(dataset_completo, 2011)

}
