% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unir_datasets.R
\name{unir_datasets}
\alias{unir_datasets}
\title{Unir datasets}
\usage{
unir_datasets(ids_estaciones)
}
\arguments{
\item{ids_estaciones}{ponemos los id de las estaciones}
}
\value{
devuelve los datasets unidos
}
\description{
Unir datasets
}
\examples{
Definir los IDs de las estaciones
ids_estaciones <- c("NH0046", "NH0098", "NH0437", "NH0472", "NH0910")
Llamar a la función para unir los datasets
dataset_completo <- unir_datasets(ids_estaciones)

}
